pyproject.toml
[project]
name = "gathering-graphrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fsspec>=2026.1.0",
    "google-genai>=1.57.0",
    "langchain-text-splitters>=1.1.0",
    "neo4j-graphrag>=1.4.0",
    "numpy>=2.4.1",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
]
